test: share tool-call promotion stream fixtures
